+++ app/src/main/java/br/com/tweetanalyzer/api/googlelanguageapi/GoogleNaturalInterface.kt
@@ -1,4 +1,4 @@
-package br.com.tweetanalyzer.GNaturalApi
+package br.com.tweetanalyzer.api.googlelanguageapi
 
 import br.com.tweetanalyzer.models.TweetAnalyseResult
 import br.com.tweetanalyzer.models.TweetAnalyseRequest
